Added Adaptor List to XmlService

+++ client/api/src/main/java/it/xeniaprogetti/cisco/ucs/plugin/client/api/UcsAdaptorEthPortErrStats.java
@@ -8,9 +8,9 @@ public class UcsAdaptorEthPortErrStats extends UcsAdaptorEthPortResourceTypeStat
        return new Builder();
     }
 
-    public final int badCrcPackets;
-    public final int badLengthPackets;
-    public final int macDiscardedPackets;
+    public final long badCrcPackets;
+    public final long badLengthPackets;
+    public final long macDiscardedPackets;
     public final String trafficDirection;
 
     private UcsAdaptorEthPortErrStats(Builder builder) {
@@ -51,9 +51,9 @@ public class UcsAdaptorEthPortErrStats extends UcsAdaptorEthPortResourceTypeStat
         public Date timeCollected;
         public long update;
 
-        private int badCrcPackets;
-        private int badLengthPackets;
-        private int macDiscardedPackets;
+        private long badCrcPackets;
+        private long badLengthPackets;
+        private long macDiscardedPackets;
         private String trafficDirection;
 
         public Builder withDn(String dn) {
@@ -86,17 +86,17 @@ public class UcsAdaptorEthPortErrStats extends UcsAdaptorEthPortResourceTypeStat
             return this;
         }
 
-        public Builder withBadCrcPackets(int badCrcPackets){
+        public Builder withBadCrcPackets(long badCrcPackets){
             this.badCrcPackets = badCrcPackets;
             return this;
         }
 
-        public Builder withBadLengthPackets(int badLengthPackets){
+        public Builder withBadLengthPackets(long badLengthPackets){
             this.badLengthPackets = badLengthPackets;
             return this;
         }
 
-        public Builder withMacDiscardedPackets(int macDiscardedPackets){
+        public Builder withMacDiscardedPackets(long macDiscardedPackets){
             this.macDiscardedPackets = macDiscardedPackets;
             return this;
         }
